+++ Source/SequenceSequencerComponent.h
@@ -27,14 +27,16 @@ class SequenceSequencerComponent   : public juce::Component
         std::vector<std::vector<int>> sliderColors={{(int)0x1001300,(int)0xFF00FF00},{(int)0x1001312,(int)0xFF000000},{(int)0x1001311,(int)0xFF00FF00},{(int)0x1001400,(int)0xFF00AF00},{(int)0x1001700,(int)0xFF000000}};
         std::vector<std::vector<int>> imageButtonColors={{(int)0x1000280,(int)0xFF00FF00},{(int)0x1000281,(int)0xFF007F00}};
 
+        juce::ComboBox sequenceSelect;
+        juce::Slider sequenceLengthSlider;
     private:
         void initSlider2(juce::Slider& slider,float min,float max,float def);
         
 
-        juce::ComboBox sequenceStep[16],sequenceSelect;
+        juce::ComboBox sequenceStep[16];
         juce::TextButton autoScroll;
         juce::Label sequenceSelectLabel,sequenceLengthSliderLabel;
-        juce::Slider sequenceLengthSlider;
+        
         
         
         JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR (SequenceSequencerComponent)