+++ HISTORY.md
@@ -2,7 +2,24 @@
 
 ## Change records
 
-### v0.03 25092023 Init project
+### v0.04 25092023 Init 
+#### TODO
+- [x] processor audio parameters definition cleaned & updated
+- [x] added listeners to GUI
+- [x] load all pluginParameters value at editor start **to be really betatested**
+- [ ] implement load/save XML presets/config (**done for audioparameters**)
+- [x] write get/Set state information for all processor parameters
+- [ ] add more effects  **repeater is a miss for now**
+- [ ] cleaner way to pass colortemplates & skin/image paths
+#### BUGS
+- [ ] some crashes in sequence mode while hiding/showing GUI
+- [ ] dry/wet for filter cracks
+- [ ] step on->off clicks for echo
+- [ ] crusher clicks 
+- [ ] buffer seems wrongly initialised on press play(sequence mode)
+- [ ] latency (same problem as choppah) -> maybe due to gain ramp
+
+### v0.03 25092023
 #### TODO
 - [x] define processor fx parameters **not all used tbc**
 - [x] put all effect in functions 
@@ -24,7 +41,7 @@
 - [ ] crusher clicks 
 - [ ] buffer seems wrongly initialised on press play(sequence mode)
 
-### v0.03 24092023 Init project
+### v0.03 24092023
 #### TODO
 - [ ] define processor fx parameters
 - [ ] load all pluginParameters value at editor start (clockDiv,fx,....)
@@ -39,7 +56,7 @@
 - [ ] cleaner way to pass colortemplates
 #### BUGS
 
-### v0.02 22092023 Init project
+### v0.02 22092023
 #### TODO
 - [x] define all processor parameters => **all but effect specific fx parameters mapped**
 - [ ] write get/Set state information for all processor parameters